code refactoring in validation logic
