+++ js/validation/validationSignIn.js
@@ -1,3 +1,5 @@
+import ifFieldValid from './validation.js'
+
 const validationSignIn = () => {
   const email = document.getElementById('email')
   const password = document.getElementById('password')
@@ -5,32 +7,6 @@ const validationSignIn = () => {
   const errorMessagePassword = document.querySelector(
     '.error-message__password'
   )
-
-  const ifFieldValid = (target, message, reg) => {
-    if (!target.value) {
-      target.classList.add('form__input_error')
-      message.innerHTML = `Please enter your ${target.type}.`
-      message.classList.add('form__text_error')
-
-      return false
-    } else {
-      target.classList.remove('form__input_error')
-      message.innerHTML = ''
-    }
-
-    if (!reg(target.value)) {
-      target.classList.add('form__input_error')
-      message.innerHTML = `Please enter a valid ${target.type}.`
-      message.classList.add('form__text_error')
-
-      return false
-    } else {
-      target.classList.remove('form__input_error')
-      message.innerHTML = ''
-    }
-
-    return true
-  }
 
   const isSignInValid =
     ifFieldValid(email, errorMessageEmail, emailIsValid) &&
