+++ js/view/signup.js
@@ -60,7 +60,7 @@ const SignUp = {
                 type="password"
                 placeholder="Confirm Password"
                 id="confirmPass"
-                autocomplete 
+                autocomplete="on"
                 required
               />
               <div class="input__svg input__svg_close"></div>
@@ -69,7 +69,7 @@ const SignUp = {
             </div>
           </div>
 
-          <button class="form__button" type="submit" data-name="/doctor">
+          <button class="form__button form__button_register" type="submit" data-name="/doctor">
             Sign up
           </button>
